EmbedTodoExtend rebuild / reddit started

+++ week08/day2/EmbedListingTodo Extend/src/main/java/com/greenfoxacademy/embedtodoextend/Repository/ToDoRepo.java
@@ -1,7 +1,11 @@
 package com.greenfoxacademy.embedtodoextend.Repository;
 import com.greenfoxacademy.embedtodoextend.Modell.ToDo;
+import org.omg.CORBA.Request;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.web.method.annotation.RequestParamMethodArgumentResolver;
 
-public interface ToDoRepo extends CrudRepository<ToDo, Integer> {
+public interface ToDoRepo extends CrudRepository<ToDo, Long> {
+
+   // ToDo saveById(Long id);
 
 }
